+++ GKGraphs/src/Implementations/LieTypeGroups/__D/__DLieTypeGroup.java
@@ -23,14 +23,14 @@ public class __DLieTypeGroup extends LieTypeGroup {
 	
 	@Override
 	protected void computeGroupOrder() {
-		int m1 = (int) (Math.pow(q, 8) + Math.pow(q, 4) + 1);
+		double m1 = Math.pow(q, 8) + Math.pow(q, 4) + 1;
 		m1 = ArithmeticUtils.primeDivMultiplication(m1);
-		int m2 = (int) (Math.pow(q, 6) - 1);
+		double m2 = Math.pow(q, 6) - 1;
 		m2 = ArithmeticUtils.primeDivMultiplication(m2);
-		int m3 = (int) (Math.pow(q, 2) - 1);
+		double m3 = Math.pow(q, 2) - 1;
 		m3 = ArithmeticUtils.primeDivMultiplication(m3);
-		this.order = p * m1 * m2 * m3;
-		MainLogger.info("*__DLieTypeGroup* Order = " + Integer.toString(order));
+		this.order = (long) (p * m1 * m2 * m3);
+		MainLogger.info("*__DLieTypeGroup* Order = " + Long.toString(order));
 	}
 	
 }
